requirements.txt
numpy
soundfile
pydub
openai-whisper
setuptools-rust